chore(cardwired): move desktop enum to it's own core file
