+++ crates/cardwire-daemon/src/analyzer/dynamic_analysis.rs
@@ -9,26 +9,7 @@ use tokio::{
     io::{AsyncBufReadExt, AsyncWriteExt, BufReader}, net::UnixStream
 };
 
-#[derive(serde::Deserialize)]
-#[serde(rename_all = "snake_case")]
-enum Desktop {
-    Niri,
-    Gnome,
-    Plasma,
-    Cosmic,
-}
-
-impl Desktop {
-    fn from_str(s: &str) -> Option<Self> {
-        match s.to_lowercase().as_str() {
-            "niri" => Some(Desktop::Niri),
-            "gnome" => Some(Desktop::Gnome),
-            "plasma" => Some(Desktop::Plasma),
-            "cosmic" => Some(Desktop::Cosmic),
-            _ => None,
-        }
-    }
-}
+use crate::core::desktop::Desktop;
 
 pub fn get_steam_app_id(environ: &[u8]) -> Option<String> {
     let prefix = b"SteamAppId=";
@@ -187,21 +168,5 @@ mod tests {
         // "CARDWIRE_ALLOW=x" value at index 15 is 'x', not '1'
         let environ = b"CARDWIRE_ALLOW=x";
         assert_eq!(check_env("CARDWIRE_ALLOW", environ), None);
-    }
-
-    #[test]
-    fn test_desktop_from_str_all_known_variants() {
-        assert!(matches!(Desktop::from_str("niri"), Some(Desktop::Niri)));
-        assert!(matches!(Desktop::from_str("gnome"), Some(Desktop::Gnome)));
-        assert!(matches!(Desktop::from_str("plasma"), Some(Desktop::Plasma)));
-        assert!(matches!(Desktop::from_str("cosmic"), Some(Desktop::Cosmic)));
-    }
-
-    #[test]
-    fn test_desktop_from_str_unknown_returns_none() {
-        assert!(Desktop::from_str("sway").is_none());
-        assert!(Desktop::from_str("hyprland").is_none());
-        assert!(Desktop::from_str("i3").is_none());
-        assert!(Desktop::from_str("").is_none());
     }
 }
